OGC Refactoring Part 11 - Bundle all OGC Distributions files in package /features/ogc_api_distirubtions
