+++ server/src/main/java/de/ingrid/igeserver/features/ogc_api_distributions/distribution_helper/OgcDistributionHelper.kt
@@ -17,7 +17,7 @@
  * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
-package de.ingrid.igeserver.features.ogcApi.distributionHelper
+package de.ingrid.igeserver.features.ogc_api_distributions.distribution_helper
 
 import com.fasterxml.jackson.databind.JsonNode
 import de.ingrid.igeserver.persistence.postgresql.jpa.model.ige.Document
