+++ server/src/main/java/de/ingrid/igeserver/features/ogc_api_distributions/profiles/bmi/distribution_helper/BmiDistributionHelper.kt
@@ -17,45 +17,41 @@
  * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
-package de.ingrid.igeserver.profiles.ingrid.distributionHelper
+package de.ingrid.igeserver.features.ogc_api_distributions.profiles.bmi.distribution_helper
 
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
-import de.ingrid.igeserver.features.ogcApi.distributionHelper.DistributionTypeInfo
-import de.ingrid.igeserver.features.ogcApi.distributionHelper.OgcDistributionHelper
+import de.ingrid.igeserver.features.ogc_api_distributions.distribution_helper.DistributionTypeInfo
+import de.ingrid.igeserver.features.ogc_api_distributions.distribution_helper.OgcDistributionHelper
 import de.ingrid.igeserver.persistence.postgresql.jpa.model.ige.Document
 import de.ingrid.igeserver.utils.getBoolean
 import de.ingrid.igeserver.utils.getString
 import de.ingrid.mdek.upload.storage.Storage
 import org.jetbrains.kotlin.utils.addToStdlib.ifFalse
-import org.springframework.context.annotation.Profile
 import org.springframework.stereotype.Service
 
-@Profile("ogc-distributions-api & ingrid")
 @Service
-class IngridDistributionHelper(
+class BmiDistributionHelper(
     private val storage: Storage
 ): OgcDistributionHelper {
 
     override val typeInfo: DistributionTypeInfo
         get() = DistributionTypeInfo(
-            "ingrid",
-            "Ingrid",
-            description = "Ingrid distribution Helper",
+            "bmi",
+            "BMI",
+            description = "BMI distribution Helper",
             emptyList()
         )
 
-    override fun canHandleDistribution(profile: String): Boolean {
-        return "ingrid" == profile
-    }
+    override fun canHandleDistribution(profile: String): Boolean = "bmi" == profile
 
     override fun getDistributionDetails(document: Document, collectionId: String, recordId: String, distributionId: String?): JsonNode {
-        val allDistributions = document.data["graphicOverviews"]
+        val allDistributions = document.data["distributions"]
 
         return if (distributionId.isNullOrEmpty()) {
             allDistributions
         } else {
-            val filteredDistributions = allDistributions.filter { it.getString("fileName.uri") == distributionId }
+            val filteredDistributions = allDistributions.filter() { it.getString("link.uri") == distributionId }
             convertListToJsonNode(filteredDistributions)
         }
     }
@@ -69,15 +65,14 @@ class IngridDistributionHelper(
     ): List<String> {
         val missingFiles: MutableList<String> = mutableListOf()
 
-        distributions.forEach { distribution ->
-            val currentDistributionId = distribution.getString("fileName.uri")!!
-            val isLink = distribution.getBoolean("fileName.asLink")!!
+        distributions.forEach() { distribution ->
+            val currentDistributionId = distribution.getString("link.uri")!!
+            val isLink = distribution.getBoolean("link.asLink")!!
             isLink.ifFalse {
                 val distributionExists = storage.exists(collectionId, userID, recordId, currentDistributionId)
                 distributionExists.ifFalse { missingFiles.add(currentDistributionId) }
             }
         }
-
         return missingFiles
     }
